Renamed StoryPackage model to ContentCollection and all affected names in all files, to be more generic to include ContentPackage also. Also generified the code and contentUri to be specific for each content collection (i.e. story package, content package).
