+++ methode-story-package-mapper.yaml
@@ -6,7 +6,7 @@ consumer:
     queueProxyHost: http://localhost:8080
     queue: kafka
     topicName: NativeCmsPublicationEvents
-    groupName: MethodeStoryPackageMapper
+    groupName: MethodeContentCollectionMapper
     backoffPeriod: 10000
     streamCount: 1
     autoCommit: true
@@ -14,7 +14,7 @@ consumer:
   healthCheck:
     name: Can connect consumer to kafka-proxy
     severity: 2
-    businessImpact: Newly published Methode Story Packages will not be available from the Content API
+    businessImpact: Newly published Methode Content Collections will not be available from the Content API
     technicalSummary: Tests that kafka-proxy is reachable and message can be consumed
     panicGuideUrl: https://dewey.ft.com/kafka-proxy.html
   systemCode: methode-web-pub
@@ -31,11 +31,11 @@ producer:
   healthCheck:
     name: Can connect producer to kafka-proxy
     severity: 2
-    businessImpact: Newly published Methode Story Packages will not be available from the Content API
+    businessImpact: Newly published Methode Content Collections will not be available from the Content API
     technicalSummary: Tests that kafka-proxy is reachable and the CmsPublicationEvents topic exists
     panicGuideUrl: https://dewey.ft.com/kafka-proxy.html
 
-contentUriPrefix: http://methode-story-package-mapper.svc.ft.com/story-package
+contentUriPrefix: http://methode-content-collection-mapper.svc.ft.com/content-collection/
 
 server:
   applicationConnectors:
