+++ src/main/java/com/ft/methodecontentcollectionmapper/exception/ContentCollectionMapperException.java
@@ -1,0 +1,10 @@
+package com.ft.methodecontentcollectionmapper.exception;
+
+public class ContentCollectionMapperException extends RuntimeException {
+    private static final long serialVersionUID = 5626181422392553276L;
+
+    public ContentCollectionMapperException(String msg, Throwable t) {
+        super(msg, t);
+    }
+
+}
