+++ src/main/java/com/ft/methodecontentcollectionmapper/exception/TransformationException.java
@@ -1,0 +1,9 @@
+package com.ft.methodecontentcollectionmapper.exception;
+
+public class TransformationException extends RuntimeException {
+    private static final long serialVersionUID = 6161690745769307558L;
+
+    public TransformationException(Throwable e) {
+        super(e);
+    }
+}
