+++ src/main/java/com/ft/methodecontentcollectionmapper/model/ContentCollectionType.java
@@ -1,0 +1,5 @@
+package com.ft.methodecontentcollectionmapper.model;
+
+public enum ContentCollectionType {
+  STORY_PACKAGE_TYPE, CONTENT_PACKAGE_TYPE
+}
