--- src/main/java/com/ft/methodestorypackagemapper/exception/StoryPackageMapperException.java
@@ -1,10 +1,0 @@
-package com.ft.methodestorypackagemapper.exception;
-
-public class StoryPackageMapperException extends RuntimeException {
-    private static final long serialVersionUID = 5626181422392553276L;
-
-    public StoryPackageMapperException(String msg, Throwable t) {
-        super(msg, t);
-    }
-
-}
