--- src/main/java/com/ft/methodestorypackagemapper/exception/TransformationException.java
@@ -1,9 +1,0 @@
-package com.ft.methodestorypackagemapper.exception;
-
-public class TransformationException extends RuntimeException {
-    private static final long serialVersionUID = 6161690745769307558L;
-
-    public TransformationException(Throwable e) {
-        super(e);
-    }
-}
